+++ src/main/java/be/baur/sds/types/BooleanType.java
@@ -1,4 +1,4 @@
-package be.baur.sds.content;
+package be.baur.sds.types;
 
 import be.baur.sds.DataType;