+++ src/main/java/be/baur/sds/types/ComparableType.java
@@ -1,4 +1,4 @@
-package be.baur.sds.content;
+package be.baur.sds.types;
 
 import java.util.Objects;
 import java.util.function.Function;