+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@
 - `Changed` Interval.from(String, Class) to from(String, Function).
 - `Changed` AbstractStringType to generic CharacterType<T>.
 - `Changed` RangedType to generic ComparableType<T>.
+- `Changed` Package .sds.content to .sds.types.
 - Added Schema.registerDataType(), and get/isDataType().
 - Added abstract Type, which AnyType and NodeType extend.
 - Added static valueOf() methods for native RangedTypes.